requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
midiutil>=1.2.1
Pillow>=8.0.0  # For PIL/Image processing 